Optionally split input stream according to JSON structure, not by newlines
Connection objects have a new field 'split' which is a function that will be
used to split up the incoming JSON objects. The default behavior is to split
by newline, but it can be switched to separate the stream by disassembling
into JSON chunks by setting this field to 'Connection.split_by_json'.

This makes bjsonrpc compatible with implementations that do not use newlines
to separate JSON objects, such as jsonrpc4j.